refactoring from TypeScript to JavaScript
