+++ client/src/index.js
@@ -1,0 +1,18 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+import { HashRouter } from "react-router-dom";
+import "bootstrap/dist/css/bootstrap.min.css";
+import "bootstrap/dist/js/bootstrap.min.js";
+import "./index.css";
+import "./App.css";
+
+ReactDOM.render(
+	<HashRouter>
+		<App />
+	</HashRouter>,
+	document.getElementById("root")
+);
+
+serviceWorker.unregister();
