--- client/src/index.tsx
@@ -1,28 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import { HashRouter } from "react-router-dom";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "bootstrap/dist/js/bootstrap.min.js";
-import "./index.css";
-import "./App.css";
-
-
-const root = ReactDOM.createRoot(document.getElementById("root")!);
-
-root.render(<HashRouter>
-  <App />
-</HashRouter>);
-
-// ReactDOM.render(
-//   <HashRouter>
-//     <App />
-//   </HashRouter>,
-//   document.getElementById("root")
-// );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
