--- client/src/types/item.type.ts
@@ -1,9 +1,0 @@
-export default interface ItemData {
-	id?: any | null;
-	item_name: string;
-	item_type: string;
-	element: string;
-	effect: string;
-	notes?: string;
-	recommended?: boolean;
-}
